requirements.txt
azure-communication-identity==1.3.1
azure-core==1.26.1
bcrypt==3.2.2
fastapi==0.88.0
jwt-signature-validator==0.0.5
pycryptodome~=3.10.1
requests==2.24.0
boto3==1.14.34
botocore==1.17.34
passlib==1.7.4
python-jose[cryptography]
passlib[bcrypt]
PyMySQL==1.0.2
python-dotenv~=0.21.0
python-multipart
PyYAML>=6.0
pyotp==2.7.0
psycopg2-binary>=2.9.9
psycopg2~=2.9.3
shortuuid==1.0.11
sqlalchemy==1.4.44
sqlalchemy_utils==0.38.3
starlette~=0.22.0
xmltodict
rsa
uvicorn~=0.13.4
configparser~=5.0.1
MarkupSafe==2.1.3
Werkzeug==3.0.1
jsondiff==2.0.0
xlsxwriter==3.1.9
numpy==1.26.1
pandas==2.1.1